Fix the observers.once() when its callable returns nil
